subscriptions/serializers.py
from rest_framework import serializers

from subscriptions.models import Subscription


class SubscriptionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Subscription
        fields = "__all__"


class EditionChoiceSerializer(serializers.Serializer):
    def update(self, instance, validated_data):
        pass

    def create(self, validated_data):
        pass

    magazine_id = serializers.IntegerField()
    time_amount = serializers.IntegerField(min_value=1)


